refactor!: add module support and take out import maps from widgets
This should be the pre-cursor to what should become ipyesm, which
could be a separate package that ipyvue, ipyreact and possibly
anywidget could depend on and use.

Note that the interface is declerative (you do not explicitly create
widgets), which allows Solara (or other frameworks that render widgets
to handle the import maps and esm loading in a different way.)

For instance, solara could add a import map DOM element, and load the
ESM modules via a script tag.